+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.1.0 (unreleased)
 - Add `Pdfkiwi.convertHtml()` method (allowing to convert html string to pdf).
 - Add `saveToFile()` utility method (allowing to save the return of `Pdfkiwi.convertHtml()` in a file).
+- Add `sendHttpResponse()` utility method (allowing to return the generated PDF in an HTTP response).